NEXT-31 Refactor main.js to multiple files
